Moved tappable_media_item.dart to media_items.dart.
